+++ reddwarf/implementations/polis.py
@@ -2,14 +2,24 @@ from typing import Optional
 from pandas import DataFrame
 from sklearn.decomposition import PCA
 from reddwarf.sklearn.cluster import PolisKMeans
+from reddwarf.types.polis import PolisRepness
 from reddwarf.utils.consensus import select_consensus_statements, ConsensusResult
-from reddwarf.utils.matrix import generate_raw_matrix, simple_filter_matrix, get_clusterable_participant_ids
+from reddwarf.utils.matrix import (
+    generate_raw_matrix,
+    simple_filter_matrix,
+    get_clusterable_participant_ids,
+)
 from reddwarf.utils.pca import run_pca
 from reddwarf.utils.clustering import find_optimal_k
 from dataclasses import dataclass
 import pandas as pd
 
-from reddwarf.utils.stats import calculate_comment_statistics_dataframes, populate_priority_calculations_into_statements_df
+from reddwarf.utils.stats import (
+    calculate_comment_statistics_dataframes,
+    populate_priority_calculations_into_statements_df,
+    select_representative_statements,
+)
+
 
 @dataclass
 class PolisClusteringResult:
@@ -25,8 +35,10 @@ class PolisClusteringResult:
         group_comment_stats (DataFrame): A multi-index dataframes for each statement, indexed by group ID and statement.
         statements_df (DataFrame): A dataframe with all intermediary and final statement data/calculations/metadata.
         participants_df (DataFrame): A dataframe with all intermediary and final participant data/calculations/metadata.
-        consensus (ConsensusResult): A dict of up to top 5 statistically statements for each of agree/disagree.
+        consensus (ConsensusResult): A dict of the most statistically significant statements for each of agree/disagree.
+        repness (PolisRepness): A dict of the most statistically significant statements most representative of each group.
     """
+
     raw_vote_matrix: DataFrame
     filtered_vote_matrix: DataFrame
     pca: PCA
@@ -38,6 +50,8 @@ class PolisClusteringResult:
     statements_df: DataFrame
     participants_df: DataFrame
     consensus: ConsensusResult
+    repness: PolisRepness
+
 
 def run_clustering(
     votes: list[dict],
@@ -49,6 +63,8 @@ def run_clustering(
     max_group_count: int = 5,
     force_group_count: Optional[int] = None,
     random_state: Optional[int] = None,
+    pick_max: int = 5,
+    confidence: float = 0.9,
 ) -> PolisClusteringResult:
     """
     An essentially feature-complete implementation of the Polis clustering algorithm.
@@ -68,6 +84,9 @@ def run_clustering(
         init_centers (list[list[float]]): Initial guesses of [x,y] coordinates for k-means (Length of list must match max_group_count)
         force_group_count (int): Instead of using silhouette scores, force a specific number of groups (k value)
         random_state (int): If set, will force determinism during k-means clustering
+        confidence (float): Percent confidence interval (in decimal), within which selected statements are deemed significant
+        pick_max (int): Max number of statements selected for consensus (per direction) and representative (per group).
+
 
     Returns:
         PolisClusteringResult: A dataclass containing clustering results, including intermediate calculations.
@@ -83,11 +102,15 @@ def run_clustering(
     # DataFrames each have "x" and "y" columns.
     participants_df, statements_df, pca = run_pca(vote_matrix=filtered_vote_matrix)
 
-    participant_ids_to_cluster = get_clusterable_participant_ids(raw_vote_matrix, vote_threshold=min_user_vote_threshold)
+    participant_ids_to_cluster = get_clusterable_participant_ids(
+        raw_vote_matrix, vote_threshold=min_user_vote_threshold
+    )
     if keep_participant_ids:
         # TODO: Make this an intersection, in case there are members of
         # keep_participant_ids list that aren't represented in vote_matrix.
-        participant_ids_to_cluster = sorted(list(set(participant_ids_to_cluster + keep_participant_ids)))
+        participant_ids_to_cluster = sorted(
+            list(set(participant_ids_to_cluster + keep_participant_ids))
+        )
 
     if force_group_count:
         k_bounds = [force_group_count, force_group_count]
@@ -105,9 +128,11 @@ def run_clustering(
     label_series = pd.Series(
         kmeans.labels_ if kmeans else None,
         index=participant_ids_to_cluster,
-        dtype="Int64", # Allows nullable/NaN values.
+        dtype="Int64",  # Allows nullable/NaN values.
     )
-    participants_df["to_cluster"] = participants_df.index.isin(participant_ids_to_cluster)
+    participants_df["to_cluster"] = participants_df.index.isin(
+        participant_ids_to_cluster
+    )
     participants_df["cluster_id"] = label_series
 
     grouped_stats_df, gac_df = calculate_comment_statistics_dataframes(
@@ -136,21 +161,31 @@ def run_clustering(
     consensus = select_consensus_statements(
         vote_matrix=raw_vote_matrix,
         mod_out_statement_ids=mod_out_statement_ids,
-        pick_n=5,
+        pick_max=pick_max,
+        confidence=confidence,
         prob_threshold=0.5,
-        confidence=0.9,
+    )
+
+    repness = select_representative_statements(
+        grouped_stats_df=grouped_stats_df,
+        mod_out_statement_ids=mod_out_statement_ids,
+        pick_max=pick_max,
+        confidence=confidence,
     )
 
     return PolisClusteringResult(
         raw_vote_matrix=raw_vote_matrix,
         filtered_vote_matrix=filtered_vote_matrix,
         pca=pca,
-        projected_participants=participants_df.loc[participant_ids_to_cluster, ["x", "y", "cluster_id"]], # deprecate?
-        projected_statements=statements_df.loc[:, ["x", "y"]], # deprecate?
+        projected_participants=participants_df.loc[
+            participant_ids_to_cluster, ["x", "y", "cluster_id"]
+        ],  # deprecate?
+        projected_statements=statements_df.loc[:, ["x", "y"]],  # deprecate?
         kmeans=kmeans,
-        group_aware_consensus=gac_df, # deprecate?
+        group_aware_consensus=gac_df,  # deprecate?
         group_comment_stats=grouped_stats_df,
         statements_df=statements_df,
         participants_df=participants_df,
         consensus=consensus,
+        repness=repness,
     )